Split ring talkback helpers
